fix(web): lazy-load stripe checkout so boot chunk stays under budget
